extract useInsightsSQLEditorOnMountCallback hook
